+++ test/goalc/test_with_game.cpp
@@ -735,6 +735,17 @@ TEST_F(WithGameTests, WeirdMultiply) {
                           "0\n"});
 }
 
+TEST_F(WithGameTests, Function128) {
+  runner.run_static_test(
+      env, testCategory, "test-function128.gc",
+      {"#<vector       1.0000       2.0000       3.0000       4.0000 @ #x400000003f800000>\n"
+       "#<vector       1.0000      20.0000       3.0000       4.0000 @ #x41a000003f800000>\n"
+       "#<vector      10.0000       2.0000       3.0000       4.0000 @ #x4000000041200000>\n"
+       " 0 1 2 3 4 5 6 7 8 9 a b c d e\n"
+       "arg0: 1 arg2: 2\n"
+       "0\n"});
+}
+
 TEST(TypeConsistency, TypeConsistency) {
   Compiler compiler;
   compiler.enable_throw_on_redefines();